Move parsing to core
